[BE][mod] todo 상품에 대한 리뷰 조회 페이지네이션 기능 ,별점 타입 변경

+++ src/main/java/com/doubleowner/revibe/domain/review/entity/Review.java
@@ -23,7 +23,7 @@ public class Review extends BaseTimeEntity {
     private Long id;
 
     @Column(nullable = false)
-    private Double starRate;
+    private Integer starRate;
 
     @Column(nullable = false)
     private String title;
